+++ src/pyuncertainnumber/pba/pbox_abc.py
@@ -951,10 +951,10 @@ def simple_stacking(itvls):
         - only meant for quick use during development
         - see `stacking` function for production use
     """
-    from .utils import weighted_ecdf, eCDF_bundle
+    from .utils import get_ecdf, eCDF_bundle
 
-    q1, p1 = weighted_ecdf(itvls.lo)
-    q2, p2 = weighted_ecdf(itvls.hi)
+    q1, p1 = get_ecdf(itvls.lo)
+    q2, p2 = get_ecdf(itvls.hi)
 
     cdf1 = eCDF_bundle(q1, p1)
     cdf2 = eCDF_bundle(q2, p2)
